loom/puppet.py - using ruby gem requires make and other build tools
The standard way of preparing a fresh Ubuntu install to build software
from source is to install the 'build-essential' package. The particular
failure without the commit is the lack of `make` when installing the
librarian-puppet module:

```
Building native extensions.  This could take a while...
ERROR:  Error installing librarian-puppet:
       ERROR: Failed to build gem native extension.

        /usr/bin/ruby1.9.1 extconf.rb
creating Makefile

make
sh: 1: make: not found
```